refactor(lightbox navigation): cleaner files
