+++ assets/javascript/modules/modals/escapeModal.js
@@ -3,9 +3,11 @@ import closeModal from './closeModal.js';
 function escapeModal(modalBackground) {
     window.addEventListener('keydown', event => {
         if (event.key === 'Escape') {
-            closeModal(modalBackground, modalBackground);
+            if (modalBackground) {
+                closeModal(modalBackground, modalBackground);
+            }
         }
-    }, {once: true});
+    });
 }
 
 export default escapeModal;
